requirements.txt
django
djangorestframework
pymysql
celery
pika
torch
torchvision
diffusers
cryptography
mysqlclient
transformers
hf_xet
accelerate
